Moved addNoise and attitudeError to conics and navigation sources files, respectively
